backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
groq>=0.4.0
pywin32>=306
psutil>=5.9.0
python-dotenv>=1.0.0
